requirements.txt
appdirs==1.4.4
cityscapesScripts==2.1.7
coloredlogs==14.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
future==0.18.2
humanfriendly==8.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pycocotools==2.0.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyquaternion==0.9.9
python-dateutil==2.8.1
six==1.15.0
timm==0.3.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.53.0
typing==3.7.4.3
typing-extensions==3.7.4.3
